Change interface to implementing class in class diagram
